+++ flake.nix
@@ -33,7 +33,7 @@
         {
           default = pkgs.sbcl.buildASDFSystem {
             pname = "nshell";
-            version = "0.3.3";
+            version = "0.4.0";
             src = src;
             systems = [ "nshell" ];
             lispLibs = [];
@@ -72,7 +72,7 @@
 
           test = pkgs.sbcl.buildASDFSystem {
             pname = "nshell-test";
-            version = "0.3.3";
+            version = "0.4.0";
             src = src;
             systems = [ "nshell/test" ];
             lispLibs = [ pkgs.sbclPackages.fiveam ];
@@ -97,7 +97,7 @@
         # Run the full test suite (332 tests)
         test = pkgs.sbcl.buildASDFSystem {
           pname = "nshell-test-check";
-          version = "0.3.3";
+          version = "0.4.0";
           src = src;
           systems = [ "nshell/test" ];
           lispLibs = [ pkgs.sbclPackages.fiveam ];